move add preview to it's own component
